requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai>=1.84.0
requests==2.31.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
python-slugify==8.0.1
markdown==3.5.1
pytest==7.4.4
APScheduler==3.10.4 